requirements.txt
discord.py==1.3.4
num2words==0.5.7
mwclient=0.10.0
nltk==3.3
PyYAML==3.13
dateparser==0.7.0
Pillow==5.3.0
lark-parser==0.7.1
isodatetime==1!2.0.0